Refactor indexer processing logic
